# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mc8yp HTTP API",
    "version": "1.0.0",
    "description": "Auxiliary REST surface of the mc8yp Cumulocity MCP microservice.\n\nThis spec intentionally documents only the non-MCP HTTP routes (`/refresh-apis`, `/resolve-mcp-servers`, `/stats`, `/health`). The MCP protocol itself is served at `POST /mcp` over JSON-RPC and is **not** documented here — MCP clients discover tools and prompts via the MCP `tools/list` / `prompts/list` flow, not via OpenAPI. If you are an agent reading this through `serviceSpecs['mc8yp-server']`, you are talking to mc8yp through mc8yp; do not call `/service/mc8yp-server/mcp` from inside `execute`.\n\nThe primary use cases for this spec are letting an agent trigger a fresh API discovery in mc8yp itself after subscribing or unsubscribing a microservice in the tenant, and letting a management UI resolve and check external MCP server configuration before storing it."
  },
  "paths": {
    "/refresh-apis": {
      "post": {
        "operationId": "refreshApis",
        "summary": "Bust the API discovery cache and re-run discovery for the calling tenant.",
        "description": "Resolves the tenant from the caller's `Authorization` header (or session cookie), busts the per-tenant discovery cache, and synchronously re-runs Cumulocity application + OpenAPI spec discovery using the microservice's service user for that tenant. Returns the freshly discovered spec metadata.\n\nUse this after a new microservice has been subscribed in the tenant — otherwise discovered specs stay cached for 30 minutes.",
        "responses": {
          "200": {
            "description": "Discovery completed successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RefreshApisResponse" }
              }
            }
          },
          "401": {
            "description": "No Authorization header or session cookie was provided.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "400": {
            "description": "Could not resolve the calling tenant via `/tenant/currentTenant`.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "403": {
            "description": "Tenant is not subscribed to this microservice.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "500": {
            "description": "Discovery failed for an unexpected reason.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/resolve-mcp-servers": {
      "post": {
        "operationId": "resolveMcpServers",
        "summary": "Resolve and check external MCP server configuration before it is used.",
        "description": "Answers the three questions a caller building the `mc8yp-mcp-server` header cannot answer itself: **which namespace** each entry would be mounted under (derived from its free-form `name` by the same sanitization a Cumulocity contextPath gets, so `MaStR registry` becomes `MaStR_registry`), **whether that namespace is available** (reserved names, namespaces already held by this tenant, and duplicates within the request), and **whether the server answers** — the MCP handshake runs with the entry's own credentials, exactly as it would at runtime, and the tool list it returns is what the agent would see.\n\nWhy this matters: at runtime a tenant namespace always wins a collision and the external server is then **silently skipped**, so the operator believes the agent has a namespace it does not have. This route turns that into a visible `namespace-taken` verdict at configuration time.\n\nThe entry shape is the header's, except that `name` may be free-form here. It is a POST with a body rather than a GET reading the header because the entries carry credentials and the candidate is typically not configured on any connection yet.\n\n**The namespace does not become dynamic.** The header contract stays strict and verbatim: resolve a display name once, store the `namespace` you got back, and send that from then on. Nothing is persisted here — mc8yp holds no external server configuration of its own.\n\nHandshakes are neither served from nor written to the per-session tool-list cache, so the answer always reflects the credentials as sent.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResolveMcpServersRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "One resolution per submitted entry, in the order they were sent. A 200 does NOT mean every entry is usable — read each `status`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ResolveMcpServersResponse" }
              }
            }
          },
          "400": {
            "description": "Body was not an object carrying a `servers` array. Per-entry problems are reported as `invalid` inside a 200, not here.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": {
            "description": "No Authorization header or session cookie was provided.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "500": {
            "description": "Resolution failed for an unexpected reason.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "operationId": "sandboxStats",
        "summary": "Point-in-time capacity/usage snapshot of the codemode sandbox runtime.",
        "description": "Returns the `@iso4/sandbox` runtime's own `stats()` snapshot: active/queued runs, warm-instance counts, memory budget and pressure state, and per-prefix instance counts. Diagnostics only — every number is stale the moment it returns, so treat it as observability, not synchronization.\n\nLike `/health`, the first call lazily initializes the shared sandbox runtime (spawning its child process) if no `codemode` call has done so yet.",
        "responses": {
          "200": {
            "description": "Snapshot taken successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SandboxStats" }
              }
            }
          },
          "500": {
            "description": "The sandbox runtime could not be initialized or did not answer.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "operationId": "health",
        "summary": "Kubernetes-style liveness/readiness probe.",
        "description": "Returns the static string `OK` with a 200 status when the process is alive. No authentication required.",
        "responses": {
          "200": {
            "description": "Process is healthy.",
            "content": {
              "text/plain": {
                "schema": { "type": "string", "const": "OK" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RefreshApisResponse": {
        "type": "object",
        "required": ["message", "tenantUrl", "tenantId", "installedContextPaths", "discovered"],
        "properties": {
          "message": {
            "type": "string",
            "const": "API spec discovery completed"
          },
          "tenantUrl": {
            "type": "string",
            "format": "uri",
            "description": "Base URL of the Cumulocity tenant the microservice is deployed against."
          },
          "tenantId": {
            "type": "string",
            "description": "Cumulocity tenant ID resolved from the caller's auth."
          },
          "installedContextPaths": {
            "type": "array",
            "description": "Context paths of every Cumulocity application currently subscribed by the tenant, regardless of whether the application exposes an OpenAPI spec.",
            "items": { "type": "string" }
          },
          "discovered": {
            "type": "array",
            "description": "One entry per application that exposes a reachable OpenAPI spec. Paths in each spec have already been prefixed with the application's service URL prefix.",
            "items": { "$ref": "#/components/schemas/DiscoveredSpecSummary" }
          }
        }
      },
      "DiscoveredSpecSummary": {
        "type": "object",
        "required": ["contextPath", "specLabel", "pathCount"],
        "properties": {
          "contextPath": {
            "type": "string",
            "description": "Cumulocity application contextPath (e.g. `dtm`)."
          },
          "specLabel": {
            "type": "string",
            "description": "Human-readable label of the spec entry. Equals the application name for single-spec applications."
          },
          "pathCount": {
            "type": "integer",
            "minimum": 0,
            "description": "Number of paths declared by the spec after the service-prefix rewrite."
          }
        }
      },
      "ResolveMcpServersRequest": {
        "type": "object",
        "required": ["servers"],
        "properties": {
          "servers": {
            "type": "array",
            "description": "Candidate entries. An empty array is valid and resolves to an empty result.",
            "items": { "$ref": "#/components/schemas/ExternalMcpServerCandidate" }
          }
        }
      },
      "ExternalMcpServerCandidate": {
        "type": "object",
        "description": "One external MCP server as an operator configures it — the `mc8yp-mcp-server` entry shape, except that `name` may be free-form because the namespace is derived from it here.",
        "required": ["name", "url"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name. Sanitized into the namespace: spaces, hyphens and dots become underscores, other non-word characters are dropped, a leading digit gains an underscore, and a JavaScript keyword gains a trailing one. A name that is already a valid identifier passes through unchanged.",
            "examples": ["MaStR registry", "github"]
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Absolute http/https MCP endpoint, used as provided."
          },
          "token": {
            "type": "string",
            "description": "Bearer shorthand: sent as `Authorization: Bearer <token>` to the external server. Not persisted and not logged."
          },
          "headers": {
            "type": "object",
            "additionalProperties": { "type": "string" },
            "description": "Extra request headers, applied after `token` so an explicit `authorization` entry replaces the bearer shorthand."
          },
          "description": {
            "type": "string",
            "description": "Shown to the agent in the namespace overview; falls back to the server's own `instructions`."
          }
        }
      },
      "ResolveMcpServersResponse": {
        "type": "object",
        "required": ["tenantUrl", "tenantNamespaces", "warnings", "servers"],
        "properties": {
          "tenantUrl": {
            "type": "string",
            "format": "uri",
            "description": "Base URL of the Cumulocity tenant the microservice is deployed against."
          },
          "tenantId": {
            "type": "string",
            "description": "Tenant resolved from the caller's auth. Absent when resolution failed."
          },
          "tenantNamespaces": {
            "type": ["array", "null"],
            "items": { "type": "string" },
            "description": "Every namespace this tenant already occupies (core plus one per available service) — the set an external namespace has to fit around. **`null` means the check did not run** (see `warnings`), so a namespace reported free may still be taken at runtime."
          },
          "warnings": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Non-fatal problems that limit how much of the answer can be trusted. Empty on a fully checked response."
          },
          "servers": {
            "type": "array",
            "description": "One entry per submitted server, same order.",
            "items": { "$ref": "#/components/schemas/ExternalMcpResolution" }
          }
        }
      },
      "ExternalMcpResolution": {
        "type": "object",
        "required": ["name", "namespace", "status"],
        "properties": {
          "name": {
            "type": "string",
            "description": "The `name` as submitted, echoed so results can be matched to input."
          },
          "namespace": {
            "type": "string",
            "description": "The namespace this entry would be mounted under — **the value to store and to send in the `mc8yp-mcp-server` header from now on**. Empty when the entry carried no usable name."
          },
          "status": {
            "type": "string",
            "enum": ["ok", "invalid", "namespace-taken", "unreachable"],
            "description": "The one field to branch on.\n- `ok` — valid, namespace free, server answered.\n- `invalid` — the entry is malformed; sending it in the header would fail the whole request with a 400.\n- `namespace-taken` — well-formed, but the namespace is claimed, so at runtime this server is skipped and the agent silently would not have it. Rename it.\n- `unreachable` — well-formed and free, but the handshake failed. mc8yp retries on the next codemode call, so this may be transient."
          },
          "reason": {
            "type": "string",
            "description": "Why the status is not `ok`, phrased for an operator. Absent when it is."
          },
          "namespaceTakenBy": {
            "type": "string",
            "enum": ["reserved", "tenant", "request"],
            "description": "Who holds the namespace: a sandbox global (`codemode`, `docs`, `sandbox`, `c8y`, `cumulocity`), a namespace of this tenant, or an earlier entry in the same request. Only set for `namespace-taken`."
          },
          "server": {
            "type": "object",
            "description": "The downstream server's own handshake identity. Present when it answered.",
            "properties": {
              "name": { "type": "string" },
              "version": { "type": "string" },
              "instructions": {
                "type": "string",
                "description": "The server's `instructions`, which mc8yp shows the agent when the entry has no `description`."
              }
            }
          },
          "tools": {
            "type": "array",
            "description": "The tools the agent would get from this namespace. Present whenever the handshake succeeded — including for `namespace-taken`, so an operator renaming the entry already knows the URL and credential are right.",
            "items": { "$ref": "#/components/schemas/ExternalMcpResolvedTool" }
          }
        }
      },
      "ExternalMcpResolvedTool": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "description": "Tool name as advertised by the server." },
          "description": { "type": "string", "description": "The tool's description, or its title when it has none." }
        }
      },
      "SandboxStats": {
        "type": "object",
        "description": "A point-in-time capacity/usage snapshot from the `@iso4/sandbox` runtime. The runtime-side numbers are mutually consistent (taken under one registry lock); `queueDepth` is a separate host-side read taken after the round trip.",
        "required": [
          "activeRuns",
          "queueDepth",
          "openConnections",
          "warmInstances",
          "idleInstances",
          "idleHeapBytes",
          "budgetBytes",
          "rssBytes",
          "usageBytes",
          "hardLineBytes",
          "underPressure",
          "prefixes"
        ],
        "properties": {
          "activeRuns": {
            "type": "integer",
            "minimum": 0,
            "description": "Runs executing right now: one-off runs plus prefix runs on warm instances. Several prefix runs may share one instance, so this can exceed the busy-instance count."
          },
          "queueDepth": {
            "type": "integer",
            "minimum": 0,
            "description": "Callers queued for a free run slot (the configured maximum of concurrent runs is already executing)."
          },
          "openConnections": {
            "type": "integer",
            "minimum": 0,
            "description": "Host-side count of runtime connections as the pool tracks them. Connections are kept for the process lifetime, so this rises to roughly the concurrency high-water mark divided by the per-connection cap and stays there."
          },
          "warmInstances": {
            "type": "integer",
            "minimum": 0,
            "description": "Resident warm instances — busy plus idle. Bounded by memory, not by a count."
          },
          "idleInstances": {
            "type": "integer",
            "minimum": 0,
            "description": "Idle warm instances ready to serve their prefix without a cold start."
          },
          "idleHeapBytes": {
            "type": "integer",
            "minimum": 0,
            "description": "Summed heap of the idle instances in bytes, each measured after its last call."
          },
          "budgetBytes": {
            "type": "integer",
            "minimum": 0,
            "description": "The memory mark the runtime sheds against, in bytes. 0 = watermarks disabled. Utilization is `usageBytes / budgetBytes`."
          },
          "rssBytes": {
            "type": "integer",
            "minimum": 0,
            "description": "The runtime child process's own resident set size in bytes at snapshot time. 0 when the platform offers no reader."
          },
          "usageBytes": {
            "type": "integer",
            "minimum": 0,
            "description": "Measured global container memory in bytes at snapshot time (cgroup working set including the Node host; equals the child's RSS where no cgroup exists; 0 when neither is readable)."
          },
          "hardLineBytes": {
            "type": "integer",
            "minimum": 0,
            "description": "The isolate admission line in bytes: 90% of the container limit minus the host reserve. A run needing a new isolate that would not fit under it fails with `ERR_CAPACITY`. 0 = no container limit readable, line disabled."
          },
          "underPressure": {
            "type": "boolean",
            "description": "True while the runtime is shedding: usage reached the budget and has not yet fallen back to 80% of it. While shedding, idle instances are evicted and new warm admissions stop."
          },
          "prefixes": {
            "type": "object",
            "description": "Per-prefix instance counts, keyed by prefix id — saturation of a single hot prefix is diagnosable here.",
            "additionalProperties": {
              "type": "object",
              "required": ["idle", "busy"],
              "properties": {
                "idle": { "type": "integer", "minimum": 0 },
                "busy": { "type": "integer", "minimum": 0 }
              }
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "description": "Standard H3 error envelope.",
        "properties": {
          "status": { "type": "integer" },
          "statusText": { "type": "string" },
          "message": { "type": "string" }
        }
      }
    }
  }
}